requirements/base.txt
python-dotenv
sqlmodel
asyncpg
nest_asyncio
argon2-cffi
greenlet
pillow
fastapi-users[oauth]
fastapi_users_db_sqlmodel